Impl std::error:Error for EnrError
